Perf improvement: Switch animated GIF to JPEG
The ~5MB animated GIF we currently ship to all users (including those
on mobile) is starting to cause some pain and should be dropped for a
static header instead.

Reasons:

1. It’s hurting jsconf.eu performance on both mobile and desktop, but
particularly the former. Without the GIF the site scores far better on
webpagetest with almost instant improvements to the speed index.

2. For international users visiting jsconf.eu on their phones and
tablets, they’re going to rack up charges they otherwise shouldn’t just
to check the site and schedule.

3. I love the GIF. GIFs should be everywhere, but only where it makes
sense. The smallest I could get the GIF we use today down to was about
3MB and for <video> instead ~600KB. Contrasting that with a far smaller
sized image of the same resolution, the site would mostly look the same
but just be a little more static.

Would you please consider this? :)